use sealed trait for osm entity
