+++ src/main/scala/edu/ie3/util/osm/OsmEntities.scala
@@ -1,0 +1,156 @@
+/*
+ * © 2021. TU Dortmund University,
+ * Institute of Energy Systems, Energy Efficiency and Energy Economics,
+ * Research group Distribution grid planning and operation
+*/
+package edu.ie3.util.osm
+
+import org.locationtech.jts.geom.Point
+
+import java.time.ZonedDateTime
+import java.util.UUID
+
+object OsmEntities {
+
+  /** Common trait to all OpenStreetMap entities
+    */
+  sealed trait OsmEntity {
+    val uuid: UUID
+    val osmId: Int
+    val lastEdited: ZonedDateTime
+    val tags: Map[String, String]
+
+    /** Checks whether the tags contain a certain key/value pair.
+      *
+      * @param key
+      *   key to look for
+      * @param value
+      *   associated value
+      * @return
+      *   Boolean
+      */
+    def containsKeyValuePair(key: String, value: String): Boolean = {
+      tags.get(key) match {
+        case Some(tagValue) => tagValue == value
+        case None           => false
+      }
+    }
+
+    /** Checks whether the tags contain a certain key that
+      *
+      * @param key
+      *   key to look for
+      * @param values
+      *   set of values to look for
+      * @return
+      *   Boolean
+      */
+    def containsKeyValuePair(key: String, values: Set[String]): Boolean = {
+      tags.get(key) match {
+        case Some(tagValue) => values.contains(tagValue)
+        case None           => false
+      }
+    }
+  }
+
+  /** Represents an OSM node
+    *
+    * @param uuid
+    *   internal uuid
+    * @param osmId
+    *   osm specific identifier
+    * @param lastEdited
+    *   time stamp of the elements last edit
+    * @param tags
+    *   associated tags
+    * @param coordinates
+    *   the coordinates of the specific node
+    */
+  final case class Node(
+      uuid: UUID,
+      osmId: Int,
+      lastEdited: ZonedDateTime,
+      tags: Map[String, String],
+      coordinates: Point
+  ) extends OsmEntity
+
+  /** Common trait to all OSM ways
+    */
+  sealed trait Way extends OsmEntity {
+    def nodes: List[Node]
+  }
+
+  /** An open way.
+    *
+    * @param uuid
+    *   internal uuid
+    * @param osmId
+    *   osm specific identifier
+    * @param lastEdited
+    *   time stamp of the elements last edit
+    * @param tags
+    *   associated tags
+    * @param nodes
+    *   nodes that make up the way
+    */
+  case class OpenWay(
+      uuid: UUID,
+      osmId: Int,
+      lastEdited: ZonedDateTime,
+      tags: Map[String, String],
+      nodes: List[Node]
+  ) extends Way
+
+  /** A closed way. A way is closed if the first node and its last node are
+    * identical.
+    *
+    * @param uuid
+    *   internal uuid
+    * @param osmId
+    *   osm specific identifier
+    * @param lastEdited
+    *   time stamp of the elements last edit
+    * @param tags
+    *   associated tags
+    * @param nodes
+    *   nodes that make up the way
+    */
+  case class ClosedWay(
+      uuid: UUID,
+      osmId: Int,
+      lastEdited: ZonedDateTime,
+      tags: Map[String, String],
+      nodes: List[Node]
+  ) extends Way
+
+  /** An OSM relation.
+    *
+    * @param uuid
+    *   internal uuid
+    * @param osmId
+    *   osm specific identifier
+    * @param lastEdited
+    *   time stamp of the elements last edit
+    * @param tags
+    *   associated tags
+    * @param elements
+    *   the elements that make up the relation
+    */
+  case class Relation(
+      uuid: UUID,
+      osmId: Int,
+      lastEdited: ZonedDateTime,
+      tags: Map[String, String],
+      elements: List[RelationElement]
+  ) extends OsmEntity
+
+  /** Represents an element of a relation.
+    *
+    * @param element
+    *   the relation element (which can either be a Way, Node or Relation)
+    * @param role
+    *   the role of the particular element
+    */
+  final case class RelationElement(element: OsmEntity, role: String)
+
+}
