--- src/main/scala/edu/ie3/util/osm/Relation.scala
@@ -1,23 +1,0 @@
-/*
- * © 2021. TU Dortmund University,
- * Institute of Energy Systems, Energy Efficiency and Energy Economics,
- * Research group Distribution grid planning and operation
-*/
-package edu.ie3.util.osm
-
-import java.time.ZonedDateTime
-import java.util.UUID
-
-/** @param uuid
-  * @param osmId
-  * @param lastEdited
-  * @param tags
-  * @param elements
-  */
-case class Relation(
-    uuid: UUID,
-    osmId: Int,
-    lastEdited: ZonedDateTime,
-    tags: Map[String, String],
-    elements: List[RelationElement]
-) extends OsmEntity
